Split out Taxon model into separate file
This commit splits out the Taxon model into a separate file for cleaner code.
